requirements.txt
phidata
python-dotenv
sqlalchemy
psycopg[binary]
pgvector
pypdf
typer
google
groq
google-generativeai